ChatHOT/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
openai
pdfplumber==0.11.5
python-dotenv==1.0.1
pydantic==2.10.6
